split messages_metadata.rs to messages_referenced.rs and messages_solidified.rs
