+++ bee-node/src/plugins/mqtt/handlers/mod.rs
@@ -18,8 +18,8 @@ pub(crate) mod addresses_ouptuts_consumed;
 pub(crate) mod addresses_ouptuts_created;
 pub(crate) mod messages;
 pub(crate) mod messages_indexation;
-pub(crate) mod messages_metadata;
 pub(crate) mod messages_referenced;
+pub(crate) mod messages_solidified;
 pub(crate) mod milestones_confirmed;
 pub(crate) mod milestones_latest;
 pub(crate) mod outputs;
